refactor: 소셜 로그인 데이터 소스 파일 경로 변경
